pyproject.toml
[tool.poetry]
name = "pythonproject8"
version = "0.1.0"
description = ""
authors = ["Mery-debug <m.a.shap58000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
setuptools = "^75.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
